notif overhaul: account for mobile
