+++ ui/src/nav/MobileGroupRoot.tsx
@@ -1,16 +1,14 @@
 import React from 'react';
-import { Link, useLocation } from 'react-router-dom';
-import { useGroupFlag, useGroup, useAmAdmin } from '@/state/groups';
-import ChannelList, { ChannelSorter } from '@/groups/GroupSidebar/ChannelList';
+import { Link } from 'react-router-dom';
+import { useGroupFlag, useGroup } from '@/state/groups';
 import GroupAvatar from '@/groups/GroupAvatar';
 import CaretLeftIcon from '@/components/icons/CaretLeft16Icon';
-import AddIcon from '@/components/icons/AddIcon';
+import Notifications from '@/notifications/Notifications';
+import GroupNotification from '@/notifications/GroupNotification';
 
 export default function MobileGroupRoot() {
   const flag = useGroupFlag();
   const group = useGroup(flag);
-  const isAdmin = useAmAdmin(flag);
-  const location = useLocation();
 
   return (
     <>
@@ -27,23 +25,9 @@ export default function MobileGroupRoot() {
             </h1>
           </Link>
         </div>
-        <div>
-          <div className="flex items-center space-x-5">
-            <ChannelSorter isMobile={true} />
-            {isAdmin && (
-              <Link
-                className="default-focus flex items-center rounded bg-blue p-1 text-base font-semibold"
-                to={`/groups/${flag}/channels/new`}
-                state={{ backgroundLocation: location }}
-              >
-                <AddIcon className="h-4 w-4 text-white" />
-              </Link>
-            )}
-          </div>
-        </div>
       </header>
-      <div className="h-full w-full flex-1 overflow-y-scroll p-2 pr-0">
-        <ChannelList flag={flag} />
+      <div className="h-full w-full flex-1 overflow-y-scroll p-0">
+        <Notifications child={GroupNotification} />
       </div>
     </>
   );
